001-prevent-race-conditions/01-blog-views/requirements.txt
fastapi
tortoise-orm[asyncpg]
uvicorn[standard]
pydantic
